test(protocol): move integration tests into a separated directory && run tests based on directory instead of grep
